Client: Profile: moved styles to separate files
