+++ packages/client/src/pages/Profile/Section/Body/sub-components/main-profile/styled-main-profile.js
