+++ packages/client/src/pages/Profile/Section/Body/sub-components/subscription/styled-subscriptions.js
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+
+export const StyledWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+
+  .toggle-btn {
+    position: relative;
+    align-items: center;
+    height: 20px;
+    grid-gap: 12px !important;
+  }
+`;
